# usoft-backend/src/docs/openapi-usof.yaml
openapi: 3.0.3
info:
  title: USOF Backend API
  version: 1.2.0
  description: |
    Forum platform backend (No ORM, pure SQL + mysql2). This spec enumerates all implemented routes, parameters, request/response schemas, and auth rules.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Category:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        slug: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, createdAt, updatedAt]
...
    Post:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        status: { type: string, enum: [active, inactive] }
        likesCount: { type: integer }
        dislikesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        authorId: { type: integer }
        authorLogin: { type: string }
        authorFullName: { type: string }
        authorRating: { type: integer }
      required: [id, title, content, status, likesCount, dislikesCount, createdAt, updatedAt]
    PostCreate:
      type: object
      properties:
        title: { type: string, minLength: 3 }
        content: { type: string, minLength: 3 }
        categories:
          type: array
          minItems: 1
          items: { type: integer }
      required: [title, content, categories]
    PostUpdate:
      type: object
      properties:
        title: { type: string, minLength: 3 }
        content: { type: string, minLength: 3 }
        categories:
          type: array
          items: { type: integer }
        status:
          type: string
          enum: [active, inactive]
    Comment:
      type: object
      properties:
        id: { type: integer }
        postId: { type: integer }
        authorId: { type: integer }
        content: { type: string }
        status: { type: string, enum: [active, inactive] }
        likesCount: { type: integer }
        dislikesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Like:
      type: object
      properties:
        id: { type: integer }
        authorId: { type: integer }
        targetType: { type: string, enum: [post, comment] }
        targetId: { type: integer }
        type: { type: string, enum: [like, dislike] }
        createdAt: { type: string, format: date-time }
    LikeCreate:
      type: object
      properties:
        type: { type: string, enum: [like, dislike], default: like }
    FavoriteOk:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    PaginatedPosts:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/Post' }
      required: [total, page, limit, items]
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /api/posts:
    get:
      tags: [Posts]
      summary: List posts (public). Visibility and status depend on role/ownership.
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
        - in: query
          name: sort
          schema: { type: string, enum: [likes, date], default: likes }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive, all], default: active }
          description: Non-admins cannot use "all" effectively; inactive visibility restricted.
        - in: query
          name: categories
          style: form
          explode: true
          schema:
            oneOf:
              - type: array
                items: { type: integer }
              - type: string
          description: IDs array, or category title string to filter by name.
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: favorite
          schema: { type: boolean }
          description: If true, returns only favorites for the current user (requires auth).
      security: []
      responses:
        '200':
          description: Paginated posts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedPosts' }

    post:
      tags: [Posts]
      summary: Create post
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostCreate' }
      responses:
        '201':
          description: Created post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/posts/{id}:
    get:
      tags: [Posts]
      summary: Get post by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Posts]
      summary: Update post (owner content/title; admin may change status; admin may NOT edit content)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostUpdate' }
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Posts]
      summary: Delete post (owner or admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/posts/{id}/comments:
    get:
      tags: [Posts, Comments]
      summary: List comments for a post
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }
  /api/posts/{id}/categories:
    get:
      tags: [Posts, Categories]
      summary: List categories for a post
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
  /api/posts/{id}/like:
    get:
      tags: [Posts]
      summary: Get likes/dislikes for a post
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Likes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Like' }

    post:
      tags: [Posts]
      summary: Like/dislike a post
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LikeCreate' }
      responses:
        '201':
          description: Created like
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Like' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Already liked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [Posts]
      summary: Remove my like/dislike from a post
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/posts/{id}/favorite:
    post:
      tags: [Posts, Favorites]
      summary: Add post to my favorites
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '201':
          description: Added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FavoriteOk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already in favorites
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [Posts, Favorites]
      summary: Remove post from my favorites
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===================== Comments ===================
...
  /api/categories/{id}/posts:
    get:
      tags: [Categories, Posts]
      summary: List posts belonging to a category
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Posts in category
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Post' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===================== Admin (REST) ===============
...
  /api/admin/posts:
    get:
      tags: [Admin]
      summary: List posts for moderation
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    title: { type: string }
                    status: { type: string, enum: [active, inactive] }
                    likesCount: { type: integer }
                    dislikesCount: { type: integer }
                    createdAt: { type: string, format: date-time }
                    updatedAt: { type: string, format: date-time }
                    authorLogin: { type: string }
                    authorId: { type: integer }
...
  /api/admin/posts/{id}:
    delete:
      tags: [Admin]
      summary: Delete post
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
...